# apps/debate-ai.com/public/debate-openapi.yml
openapi: 3.1.0
info:
  title: Debate AI API
  description: API for the Debate AI platform — debate research, video archives, leaderboards, AI analysis, and reference data.
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API routes
paths:
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      summary: Get debate team rankings and leaderboard
      description: |
        Returns ranked leaderboard data for a given division and year.
        For the current year (2026), merges TOC bid list data with DebateDrills Elo ratings.
        For prior years (2021–2025), returns DebateDrills Elo data only (VPF/VLD).
      parameters:
        - name: year
          in: query
          schema:
            type: string
            default: "2026"
          description: The competition year to query.
        - name: division
          in: query
          schema:
            type: string
            default: VPF
            enum: [VCX, VPF, VLD]
          description: The debate division (Varsity CX, Varsity PF, Varsity LD).
      responses:
        "200":
          description: Array of leaderboard entries, optionally with Elo ratings merged in.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "400":
          description: Invalid division specified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Server error fetching leaderboard data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorWithDetails"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ErrorWithDetails:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
...
    LeaderboardEntry:
      type: object
      properties:
        teamName:
          type: string
          description: School and team code (e.g., "Strake Jesuit MS").
        bids:
          type: integer
          description: Number of TOC qualifying bids earned.
        students:
          type: string
          description: Student name(s) on the team.
        debateElo:
          type: number
          description: DebateDrills Elo rating (when available).
      required:
        - teamName
...